add details

+++ application/models/product_project_model.php
@@ -69,7 +69,11 @@ class Product_project_model extends CI_Model {
 	public function deleteById($id) {
 		if ($this->db->delete('product_type', array('project' => $id))) {
 			if ($this->db->delete('product_inner', array('project' => $id))) {
-				return $this->db->delete(self::DB_NAME, array('id' => $id));
+				if ($this->db->delete('product_details', array('project' => $id))) {
+					return $this->db->delete(self::DB_NAME, array('id' => $id));
+				} else {
+					return FALSE;
+				}
 			} else {
 				return FALSE;
 			}
